Hardware Correction
A standardised board is taking shape. The pin allocations within this file differed from the actual draft board. Also, the IRQ pin of the nRF24 was connected to B5, which drives the LED used for flashes. The board was modified to connect the nRF24 IRQ to B4 thereby avoiding any contention.